Fix PlayerDamageHandler to properly deal damage and allow very small amounts of damage.

+++ Assets/Scripts/Boss/Abilities/BigSlowPellet/BossAbilityBigSlowPelletCollision.cs
@@ -5,7 +5,7 @@ using Mirror;
 
 public class BossAbilityBigSlowPelletCollision : NetworkBehaviour
 {
-    int DAMAGE = 10;
+    float DAMAGE = 10;
 
     void OnTriggerEnter2D(Collider2D playerCollider)
     {
