+++ Assets/Scripts/Boss/BossModifications.cs
@@ -6,7 +6,7 @@ using Mirror;
 public class BossModifications : NetworkBehaviour
 {
     [SyncVar]
-    public int damageDealtModPercent = 100; // ex: at 110 will do 10% more DMG
+    public float damageDealtModPercent = 100; // ex: at 110 will do 10% more DMG
 
     // Start is called before the first frame update
     void Start()
